@@ -1,33 +1,35 @@
 {
   "name": "@microsoft/gulp-core-build-webpack",
-  "version": "3.0.6",
+  "version": "3.4.108",
   "description": "",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
+  "tsdoc": {
+    "tsdocFlavor": "AEDoc"
+  },
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/microsoft/web-build-tools"
+    "url": "https://github.com/Microsoft/web-build-tools/tree/master/core-build/gulp-core-build-webpack"
   },
   "scripts": {
-    "build": "gulp",
-    "clean": "gulp clean",
-    "test": "gulp"
+    "build": "gulp --clean"
   },
   "dependencies": {
-    "@microsoft/gulp-core-build": "3.0.6",
+    "@microsoft/gulp-core-build": "3.9.26",
     "@types/gulp": "3.8.32",
-    "@types/node": "6.0.62",
+    "@types/node": "8.5.8",
+    "colors": "~1.2.1",
     "gulp": "~3.9.1",
-    "gulp-util": "~3.0.7",
-    "webpack": "~3.5.5"
+    "webpack": "~3.11.0"
   },
   "devDependencies": {
-    "@microsoft/node-library-build": "~3.2.0",
+    "@microsoft/rush-stack-compiler-3.2": "0.3.19",
+    "@microsoft/node-library-build": "6.0.67",
     "@types/orchestrator": "0.0.30",
     "@types/q": "0.0.32",
     "@types/source-map": "0.5.0",
     "@types/uglify-js": "2.6.29",
-    "@types/webpack": "3.0.10"
+    "@types/webpack": "4.4.0"
   }
 }
