@@ -1,234 +1,70 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import * as fsx from 'fs-extra';
-import * as Gulp from 'gulp';
-import * as os from 'os';
 import * as path from 'path';
-import * as through from 'through2';
-import * as gulpUtil from 'gulp-util';
-import { GulpTask } from '@microsoft/gulp-core-build';
+import { IBuildConfig } from '@microsoft/gulp-core-build';
 import {
-  Extractor,
-  IExtractorOptions,
-  IExtractorAnalyzeOptions,
-  ApiFileGenerator,
-  ApiJsonGenerator
-} from '@microsoft/api-extractor';
-import { TypeScriptConfiguration } from './TypeScriptConfiguration';
-import ts = require('gulp-typescript');
+  JsonFile,
+  FileSystem
+} from '@microsoft/node-core-library';
+import { ExtractorConfig, IExtractorInvokeOptions } from '@microsoft/api-extractor';
+import { ApiExtractorRunner as TApiExtractorRunner } from '@microsoft/rush-stack-compiler-3.1';
 
-function writeStringToGulpUtilFile(content: string, filename: string = 'tempfile'): gulpUtil.File {
-  return new gulpUtil.File({
-    contents: new Buffer(content),
-    path: filename
-  });
-}
+import { RSCTask, IRSCTaskConfig } from './RSCTask';
 
 /** @public */
-export interface IApiExtractorTaskConfig {
-  /**
-   * Indicates whether the task should be run.
-   */
-  enabled?: boolean;
-
-  /**
-   * The file path of the exported entry point, relative to the project folder.
-   *
-   * Example "src/index.ts"
-   */
-  entry?: string;
-
-  /**
-   * The file path of the folder containing API files to be reviewed, relative to
-   * the project folder.  This is part of an API review workflow:  During a build,
-   * the ApiExtractorTask will output an API file, e.g. "my-project/temp/my-project.api.ts".
-   * It will then compare this file against the last reviewed file,
-   * e.g. "../api-review/my-project.api.ts" (assuming that apiReviewFolder is "../api-review").
-   * If the files are different, the build will fail with an error message that instructs
-   * the developer to update the approved file, and then commit it to Git.  When they
-   * create a Pull Request, a VSO branch policy will look for changes under "api-review/*"
-   * and require signoff from the appropriate reviewers.
-   *
-   * Example: "config" (for a standalone project)
-   * Example: "../../common/api-review"  (for a Git repository with Rush)
-   */
-  apiReviewFolder?: string;
-
-  /**
-   * The file path of the folder containing the *.api.json output file containing
-   * the API information. The default location is in the “dist” folder,
-   * e.g. my-project/dist/my-project.api.json. This file should be published as part
-   * of the NPM package. When building other projects that depend on this package,
-   * api-extractor will look for this file in the node_modules folder and use it as an input.
-   * The *.api.json file is also consumed by a tool at
-   * https://github.com/SharePoint/ts-spec-gen that generates an online API documentation.
-   */
-  apiJsonFolder?: string;
+export interface IApiExtractorTaskConfig extends IRSCTaskConfig {
 }
 
 /**
  * The ApiExtractorTask uses the api-extractor tool to analyze a project for public APIs. api-extractor will detect
  * common problems and generate a report of the exported public API. The task uses the entry point of a project to
  * find the aliased exports of the project. An api-extractor.ts file is generated for the project in the temp folder.
- * @public
+ * @beta
  */
-export class ApiExtractorTask extends GulpTask<IApiExtractorTaskConfig>  {
+export class ApiExtractorTask extends RSCTask<IApiExtractorTaskConfig>  {
   constructor() {
     super(
       'api-extractor',
-      {
-        enabled: false,
-        entry: undefined,
-        apiReviewFolder: undefined,
-        apiJsonFolder: undefined
-      }
+      {}
     );
   }
 
   public loadSchema(): Object {
-    return require('./schemas/api-extractor.schema.json');
+    return JsonFile.load(path.resolve(__dirname, 'schemas', 'api-extractor.schema.json'));
   }
 
-  public executeTask(gulp: typeof Gulp, completeCallback: (error?: string) => void): NodeJS.ReadWriteStream | void {
-    if (!this.taskConfig.enabled || !this._validateConfiguration()) {
-      completeCallback();
-      return;
-    }
+  public isEnabled(buildConfig: IBuildConfig): boolean {
+    return FileSystem.exists(this._getApiExtractorConfigFilePath(buildConfig.rootPath));
+  }
 
-    if (!this.taskConfig.entry) {
-      completeCallback('taskConfig.entry must be defined');
-      return;
-    }
+  public executeTask(): Promise<void> {
+    this.initializeRushStackCompiler();
 
-    if (!this.taskConfig.apiJsonFolder) {
-      completeCallback('taskConfig.apiJsonFolder must be defined');
-      return;
-    }
-
-    if (!this.taskConfig.apiReviewFolder) {
-      completeCallback('taskConfig.apiReviewFolder must be defined');
-      return;
-    }
-
-    const entryPointFile: string = path.join(this.buildConfig.rootPath, this.taskConfig.entry);
-    const typingsFilePath: string = path.join(this.buildConfig.rootPath, 'typings/tsd.d.ts');
-    const otherFiles: string[] = fsx.existsSync(typingsFilePath) ? [typingsFilePath] : [];
-
-    // tslint:disable-next-line:no-any
-    const compilerOptions: ts.Settings =
-      TypeScriptConfiguration.getGulpTypescriptOptions(this.buildConfig).compilerOptions;
-
-    if (compilerOptions.module !== 'commonjs' && compilerOptions.module) {
-      this.logWarning(`Your tsconfig.json file specifies a different "target" than expected. `
-        + `Expected: "commonjs". Actual: "${compilerOptions.module}". Using "commonjs" instead.`);
-      compilerOptions.module = 'commonjs';
-    }
-
-    const extractorOptions: IExtractorOptions = {
-      compilerOptions: ts.createProject(compilerOptions).options,
-      errorHandler: (message: string, fileName: string, lineNumber: number): void => {
-        this.logWarning(`${message}` + os.EOL
-          + `  ${fileName}#${lineNumber}`);
-      }
+    const extractorOptions: IExtractorInvokeOptions = {
+      localBuild: !this.buildConfig.production
     };
 
-    const analyzeOptions: IExtractorAnalyzeOptions = {
-      entryPointFile,
-      otherFiles
-    } as any; /* tslint:disable-line:no-any */
+    const ApiExtractorRunner: typeof TApiExtractorRunner = this._rushStackCompiler.ApiExtractorRunner;
+    const extractorConfig: ExtractorConfig = ApiExtractorRunner.apiExtractor.ExtractorConfig.loadFileAndPrepare(
+      this._getApiExtractorConfigFilePath(this.buildConfig.rootPath)
+    );
 
-    const extractor: Extractor = new Extractor(extractorOptions);
-    extractor.loadExternalPackages(path.join(__dirname, 'external-api-json'));
-    extractor.analyze(analyzeOptions);
+    const apiExtractorRunner: TApiExtractorRunner = new ApiExtractorRunner(
+      extractorConfig,
+      extractorOptions,
+      this.buildFolder,
+      this._terminalProvider
+    );
 
-    const jsonGenerator: ApiJsonGenerator = new ApiJsonGenerator();
-    // const jsonContent: string = generator.generateJsonFileContent(analyzer);
-    const jsonFileName: string = path.basename(this.buildConfig.rootPath) + '.api.json';
-
-    if (!fsx.existsSync(this.taskConfig.apiJsonFolder)) {
-      fsx.mkdirsSync(this.taskConfig.apiJsonFolder, (err) => {
-        if (err) {
-          this.logError(`Could not create directory ${this.taskConfig.apiJsonFolder}`);
-        }
-      });
-    }
-
-    if (fsx.existsSync(this.taskConfig.apiJsonFolder)) {
-      const jsonFilePath: string = path.join(this.taskConfig.apiJsonFolder, jsonFileName);
-      this.logVerbose(`Writing Api JSON file to ${jsonFilePath}`);
-      jsonGenerator.writeJsonFile(jsonFilePath, extractor);
-    }
-
-    const generator: ApiFileGenerator = new ApiFileGenerator();
-    const actualApiFileContent: string = generator.generateApiFileContent(extractor);
-
-    // Ex: "project.api.ts"
-    const apiFileName: string = path.basename(this.buildConfig.rootPath) + '.api.ts';
-    this.logVerbose(`Output filename is "${apiFileName}"`);
-
-    const actualApiFilePath: string = path.join(this.buildConfig.tempFolder, apiFileName);
-
-    let foundSourceFiles: number = 0;
-    const self: ApiExtractorTask = this;
-    const expectedApiFilePath: string = path.join(this.taskConfig.apiReviewFolder, apiFileName);
-    return gulp.src(expectedApiFilePath)
-      /* tslint:disable-next-line:no-function-expression */
-      .pipe(through.obj(function (file: gulpUtil.File, enc: string, callback: () => void): void {
-        const expectedApiFileContent: string = (file.contents as Buffer).toString(enc);
-        foundSourceFiles++;
-
-        if (!ApiFileGenerator.areEquivalentApiFileContents(actualApiFileContent, expectedApiFileContent)) {
-          if (self.buildConfig.production) {
-            // For production, issue a warning that will break the CI build.
-            self.logWarning('You have changed the Public API signature for this project.  Please overwrite '
-              // @microsoft/gulp-core-build seems to run JSON.stringify() on the error messages for some reason,
-              // so try to avoid escaped characters:
-              + `'${expectedApiFilePath.replace(/\\/g, '/')}' with a copy of '${actualApiFilePath.replace(/\\/g, '/')}'`
-              + ' and then request an API review. See the Git repository README.md for more info.');
-          } else {
-            // For a local build, just copy the file automatically.
-            self.log('You have changed the Public API signature for this project.  Updating '
-              + `'${expectedApiFilePath}'`);
-            fsx.writeFileSync(expectedApiFilePath, actualApiFileContent);
-          }
-        }
-
-        callback();
-      }, function (callback: () => void): void {
-        if (foundSourceFiles === 0) {
-          // NOTE: This warning seems like a nuisance, but it has caught genuine mistakes.
-          // For example, when projects were moved into category folders, the relative path for
-          // the API review files ended up in the wrong place.
-          self.logError(`This file is missing from the "apiReviewFolder": "${expectedApiFilePath}"`
-            + ` Please copy it from the project's "temp" folder and commit it.`);
-        } else if (foundSourceFiles > 1) {
-          self.logError(`More than one file matching "${expectedApiFilePath}" was found. This is not expected.`);
-        }
-
-        this.push(writeStringToGulpUtilFile(actualApiFileContent, apiFileName));
-
-        callback();
-      }))
-      .pipe(gulp.dest(this.buildConfig.tempFolder))
-      .on('finish', () => completeCallback());
+    return apiExtractorRunner.invoke();
   }
 
-  private _validateConfiguration(): boolean {
-    if (!this.taskConfig.entry) {
-      this.logError('Missing or empty "entry" field in api-extractor.json');
-      return false;
-    }
-    if (!this.taskConfig.apiReviewFolder) {
-      this.logError('Missing or empty "apiReviewFolder" field in api-extractor.json');
-      return false;
-    }
+  protected _getConfigFilePath(): string {
+    return path.join('.', 'config', 'gcb-api-extractor.json'); // There aren't config options specific to this task
+  }
 
-    if (!fsx.existsSync(this.taskConfig.entry)) {
-      this.logError(`Entry file ${this.taskConfig.entry} does not exist.`);
-      return false;
-    }
-
-    return true;
+  private _getApiExtractorConfigFilePath(rootPath: string): string {
+    return path.resolve(rootPath, 'config', 'api-extractor.json');
   }
 }
