@@ -1,24 +1,24 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import Utilities from './Utilities';
+import { Utilities } from './Utilities';
 import * as semver from 'semver';
 
-/**
- * @public
- */
-export default class Npm {
+export class Npm {
   public static publishedVersions(
     packageName: string,
     cwd: string,
-    env: { [key: string]: string }
+    env: { [key: string]: string | undefined },
+    extraArgs: string[] = []
   ): string[] {
     const versions: string[] = [];
     try {
       const packageTime: string = Utilities.executeCommandAndCaptureOutput('npm',
-        `view ${packageName} time --json`.split(' '),
+        ['view', packageName, 'time', '--json', ...extraArgs],
         cwd,
-        env);
+        env,
+        true
+      );
       if (packageTime && packageTime !== '') {
         Object.keys(JSON.parse(packageTime)).forEach(v => {
           if (semver.valid(v)) {
@@ -29,9 +29,11 @@ export default class Npm {
         console.log(`Package ${packageName} time value does not exist. Fall back to versions.`);
         // time property does not exist. It happens sometimes. Fall back to versions.
         const packageVersions: string = Utilities.executeCommandAndCaptureOutput('npm',
-          `view ${packageName} versions --json`.split(' '),
+          ['view', packageName, 'versions', '--json', ...extraArgs],
           cwd,
-          env);
+          env,
+          true
+        );
         if (packageVersions && packageVersions.length > 0) {
           (JSON.parse(packageVersions)).forEach(v => {
             versions.push(v);
@@ -44,7 +46,7 @@ export default class Npm {
       if (error.message.indexOf('npm ERR! 404') >= 0) {
         console.log(`Package ${packageName} does not exist in the registry.`);
       } else {
-        console.log(`Failed to get npm information about ${packageName}.`);
+        console.log(`Failed to get NPM information about ${packageName}.`);
         throw error;
       }
     }
