@@ -2,31 +2,117 @@
 // See LICENSE in the project root for license information.
 
 import * as child_process from 'child_process';
-import * as fsx from 'fs-extra';
+import * as fs from 'fs';
 import * as os from 'os';
-import * as rimraf from 'rimraf';
 import * as tty from 'tty';
+import * as path from 'path';
+import * as wordwrap from 'wordwrap';
+import {
+  JsonFile,
+  IPackageJson,
+  FileSystem,
+  FileConstants
+} from '@microsoft/node-core-library';
+import { RushConfiguration } from '../api/RushConfiguration';
+
+export interface IEnvironment {
+  // NOTE: the process.env doesn't actually support "undefined" as a value.
+  // If you try to assign it, it will be converted to the text string "undefined".
+  // But this typing is needed for reading values from the dictionary, and for
+  // subsets that get combined.
+  [environmentVariableName: string]: string | undefined;
+}
 
 /**
- * @public
+ * Options for Utilities.installPackageInDirectory().
  */
-export default class Utilities {
+export interface IInstallPackageInDirectoryOptions {
+  directory: string;
+  packageName: string;
+  version: string;
+  tempPackageTitle: string;
+  maxInstallAttempts: number;
+  commonRushConfigFolder: string | undefined;
+  suppressOutput?: boolean;
+}
+
+export interface ILifecycleCommandOptions {
   /**
-   * For a scoped NPM package name this separates the scope and name parts.  For example:
-   * parseScopedPackageName('@my-scope/myproject') = { scope: '@my-scope', name: 'myproject' }
-   * parseScopedPackageName('myproject') = { scope: '', name: 'myproject' }
+   * The rush configuration, if the command is running in a rush repo.
    */
-  public static parseScopedPackageName(scopedName: string): { scope: string, name: string } {
-    if (scopedName.substr(0, 1) !== '@') {
-      return { scope: '', name: scopedName };
+  rushConfiguration: RushConfiguration | undefined;
+
+  /**
+   * Working directory for running the command
+   */
+  workingDirectory: string;
+
+  /**
+   * The folder containing a local .npmrc, which will be used for the INIT_CWD environment variable
+   */
+  initCwd: string;
+
+  /**
+   * If true, suppress the process's output, but if there is a nonzero exit code then print stderr
+   */
+  handleOutput: boolean;
+
+  /**
+   * Options for what should be added to the PATH variable
+   */
+  environmentPathOptions: IEnvironmentPathOptions;
+}
+
+export interface IEnvironmentPathOptions {
+  /**
+   * If true, include <project root>/node_modules/.bin in the PATH. If both this and
+   * {@link IEnvironmentPathOptions.includeRepoBin} are set, this path will take precedence.
+   */
+  includeProjectBin?: boolean;
+
+  /**
+   * If true, include <repo root>/common/temp/node_modules/.bin in the PATH.
+   */
+  includeRepoBin?: boolean;
+}
+
+interface ICreateEnvironmentForRushCommandPathOptions extends IEnvironmentPathOptions {
+  projectRoot: string | undefined;
+  commonTempFolder: string | undefined;
+}
+
+interface ICreateEnvironmentForRushCommandOptions {
+  /**
+   * The INIT_CWD environment variable
+   */
+  initCwd?: string;
+
+  /**
+   * an existing environment to copy instead of process.env
+   */
+  initialEnvironment?: IEnvironment;
+
+  /**
+   * Options for what should be added to the PATH variable
+   */
+  pathOptions?: ICreateEnvironmentForRushCommandPathOptions;
+}
+
+export class Utilities {
+  /**
+   * Get the user's home directory. On windows this looks something like "C:\users\username\" and on UNIX
+   * this looks something like "/usr/username/"
+   */
+  public static getHomeDirectory(): string {
+    const unresolvedUserFolder: string | undefined = process.env[
+      (process.platform === 'win32') ? 'USERPROFILE' : 'HOME'
+    ];
+    const homeFolder: string = path.resolve(unresolvedUserFolder);
+    if (!FileSystem.exists(homeFolder)) {
+      throw new Error('Unable to determine the current user\'s home directory');
     }
 
-    const slashIndex: number = scopedName.indexOf('/');
-    if (slashIndex >= 0) {
-      return { scope: scopedName.substr(0, slashIndex), name: scopedName.substr(slashIndex + 1) };
-    } else {
-      throw new Error('Invalid scoped name: ' + scopedName);
-    }
+    return homeFolder;
   }
 
   /**
@@ -55,10 +141,12 @@ export default class Utilities {
    * Retries a function until a timeout is reached. The function is expected to throw if it failed and
    *  should be retried.
    */
-  public static retryUntilTimeout<TResult>(fn: () => TResult,
-                                           maxWaitTimeMs: number,
-                                           getTimeoutError: (innerError: Error) => Error,
-                                           fnName: string): TResult {
+  public static retryUntilTimeout<TResult>(
+    fn: () => TResult,
+    maxWaitTimeMs: number,
+    getTimeoutError: (innerError: Error) => Error,
+    fnName: string
+  ): TResult {
     const startTime: number = Utilities.getTimeInMs();
     let looped: boolean = false;
 
@@ -80,6 +168,8 @@ export default class Utilities {
     if (looped) {
       const currentTime: number = Utilities.getTimeInMs();
       const totalSeconds: string = ((currentTime - startTime) / 1000.0).toFixed(2);
+      // This logging statement isn't meaningful to the end-user. `fnName` should be updated
+      // to something like `operationDescription`
       console.log(`${fnName}() stalled for ${totalSeconds} seconds`);
     }
 
@@ -87,7 +177,7 @@ export default class Utilities {
   }
 
   /**
-   * Creates the specified folder by calling fsx.mkdirsSync(), but using a
+   * Creates the specified folder by calling FileSystem.ensureFolder(), but using a
    * retry loop to recover from temporary locks that may be held by other processes.
    * If the folder already exists, no error occurs.
    */
@@ -98,29 +188,33 @@ export default class Utilities {
       return;
     }
 
-    // We need to do a simple "fs.mkdirSync(localModulesFolder)" here,
+    // We need to do a simple "FileSystem.ensureFolder(localModulesFolder)" here,
     // however if the folder we deleted above happened to contain any files,
     // then there seems to be some OS process (virus scanner?) that holds
     // a lock on the folder for a split second, which causes mkdirSync to
     // fail.  To workaround that, retry for up to 7 seconds before giving up.
     const maxWaitTimeMs: number = 7 * 1000;
 
-    return Utilities.retryUntilTimeout(() => fsx.mkdirsSync(folderName),
-                                       maxWaitTimeMs,
-                                       (e) => new Error(`Error: ${e}${os.EOL}Often this is caused by a file lock ` +
-                                                        'from a process such as your text editor, command prompt, ' +
-                                                        'or "gulp serve"'),
-                                       'createFolderWithRetry');
+    return Utilities.retryUntilTimeout(
+      () => FileSystem.ensureFolder(folderName),
+      maxWaitTimeMs,
+      (e) => new Error(
+        `Error: ${e}${os.EOL}Often this is caused by a file lock ` +
+        'from a process such as your text editor, command prompt, ' +
+        'or "gulp serve"'
+      ),
+      'createFolderWithRetry'
+    );
   }
 
   /**
    * Determines if the path points to a file and that it exists.
    */
-  public static fileExists(path: string): boolean {
+  public static fileExists(filePath: string): boolean {
     let exists: boolean = false;
 
     try {
-      const lstat: fsx.Stats = fsx.lstatSync(path);
+      const lstat: fs.Stats = FileSystem.getLinkStatistics(filePath);
       exists = lstat.isFile();
     } catch (e) { /* no-op */ }
 
@@ -130,11 +224,11 @@ export default class Utilities {
   /**
    * Determines if a path points to a directory and that it exists.
    */
-  public static directoryExists(path: string): boolean {
+  public static directoryExists(directoryPath: string): boolean {
     let exists: boolean = false;
 
     try {
-      const lstat: fsx.Stats = fsx.lstatSync(path);
+      const lstat: fs.Stats = FileSystem.getLinkStatistics(directoryPath);
       exists = lstat.isDirectory();
     } catch (e) { /* no-op */ }
 
@@ -148,7 +242,7 @@ export default class Utilities {
    */
   public static dangerouslyDeletePath(folderPath: string): void {
     try {
-      rimraf.sync(folderPath, { disableGlob: true });
+      FileSystem.deleteFolder(folderPath);
     } catch (e) {
       throw new Error(e.message + os.EOL + 'Often this is caused by a file lock'
         + ' from a process such as your text editor, command prompt, or "gulp serve"');
@@ -161,7 +255,7 @@ export default class Utilities {
   public static deleteFile(filePath: string): void {
     if (Utilities.fileExists(filePath)) {
       console.log(`Deleting: ${filePath}`);
-      fsx.unlinkSync(filePath);
+      FileSystem.deleteFile(filePath);
     }
   }
 
@@ -173,17 +267,17 @@ export default class Utilities {
    * timestamp is compared.
    */
   public static isFileTimestampCurrent(outputFilename: string, inputFilenames: string[]): boolean {
-    if (!fsx.existsSync(outputFilename)) {
+    if (!FileSystem.exists(outputFilename)) {
       return false;
     }
-    const outputStats: fsx.Stats = fsx.statSync(outputFilename);
+    const outputStats: fs.Stats = FileSystem.getStatistics(outputFilename);
 
     for (const inputFilename of inputFilenames) {
-      if (!fsx.existsSync(inputFilename)) {
+      if (!FileSystem.exists(inputFilename)) {
         return false;
       }
 
-      const inputStats: fsx.Stats = fsx.statSync(inputFilename);
+      const inputStats: fs.Stats = FileSystem.getStatistics(inputFilename);
       if (outputStats.mtime < inputStats.mtime) {
         return false;
       }
@@ -200,7 +294,24 @@ export default class Utilities {
     if (stdout && stdout.columns) {
       return stdout.columns;
     }
+
     return 80;
+  }
+
+  /**
+   * Applies word wrapping.  If maxLineLength is unspecified, then it defaults to the console
+   * width.
+   */
+  public static wrapWords(text: string, maxLineLength?: number, indent?: number): string {
+    if (!indent) {
+      indent = 0;
+    }
+    if (!maxLineLength) {
+      maxLineLength = Utilities.getConsoleWidth();
+    }
+
+    const wrap: (textToWrap: string) => string = wordwrap.soft(indent, maxLineLength);
+    return wrap(text);
   }
 
   /**
@@ -208,11 +319,15 @@ export default class Utilities {
    * The current directory will be set to the specified workingDirectory.
    */
   public static executeCommand(command: string, args: string[], workingDirectory: string,
-    suppressOutput: boolean = false, environmentVariables?: { [key: string]: string }): void {
+    environment?: IEnvironment, suppressOutput: boolean = false,
+    keepEnvironment: boolean = false
+  ): void {
 
     Utilities._executeCommandInternal(command, args, workingDirectory,
       suppressOutput ? undefined : [0, 1, 2],
-      environmentVariables);
+      environment,
+      keepEnvironment
+    );
   }
 
   /**
@@ -220,11 +335,18 @@ export default class Utilities {
    * The current directory will be set to the specified workingDirectory.
    */
   public static executeCommandAndCaptureOutput(command: string, args: string[], workingDirectory: string,
-    environmentVariables?: { [key: string]: string }): string {
+    environment?: IEnvironment,
+    keepEnvironment: boolean = false
+  ): string {
 
-    const  result: child_process.SpawnSyncReturns<Buffer>
-      = Utilities._executeCommandInternal(command, args, workingDirectory,
-        ['pipe', 'pipe', 'pipe'], environmentVariables);
+    const  result: child_process.SpawnSyncReturns<Buffer> = Utilities._executeCommandInternal(
+      command,
+      args,
+      workingDirectory,
+      ['pipe', 'pipe', 'pipe'],
+      environment,
+      keepEnvironment
+    );
 
     return result.stdout.toString();
   }
@@ -232,8 +354,9 @@ export default class Utilities {
   /**
    * Attempts to run Utilities.executeCommand() up to maxAttempts times before giving up.
    */
-  public static executeCommandWithRetry(command: string, args: string[], maxAttempts: number,
-    workingDirectory: string, suppressOutput: boolean = false): void {
+  public static executeCommandWithRetry(maxAttempts: number, command: string, args: string[],
+    workingDirectory: string,  environment?: IEnvironment, suppressOutput: boolean = false,
+    retryCallback?: () => void): void {
 
     if (maxAttempts < 1) {
       throw new Error('The maxAttempts parameter cannot be less than 1');
@@ -244,7 +367,7 @@ export default class Utilities {
     // tslint:disable-next-line:no-constant-condition
     while (true) {
       try {
-        Utilities.executeCommand(command, args, workingDirectory, suppressOutput);
+        Utilities.executeCommand(command, args, workingDirectory, environment, suppressOutput);
       } catch (error) {
         console.log(os.EOL + 'The command failed:');
         console.log(` ${command} ` + args.join(' '));
@@ -253,118 +376,56 @@ export default class Utilities {
         if (attemptNumber < maxAttempts) {
           ++attemptNumber;
           console.log(`Trying again (attempt #${attemptNumber})...` + os.EOL);
+          if (retryCallback) {
+            retryCallback();
+          }
+
           continue;
         } else {
           console.error(`Giving up after ${attemptNumber} attempts` + os.EOL);
           throw error;
         }
       }
+
       break;
     }
   }
 
   /**
-   * Executes the command with the specified command-line parameters, and waits for it to complete.
-   * The current directory will be set to the specified workingDirectory.
+   * Executes the command using cmd if running on windows, or using sh if running on a non-windows OS.
+   * @param command - the command to run on shell
+   * @param options - options for how the command should be run
    */
-  public static executeCommandAsync(command: string, args: string[], workingDirectory: string,
-    environmentVariables?: { [key: string]: string }): child_process.ChildProcess {
-    // This is a workaround for GitHub issue #25330.  It is not as complete as the workaround above,
-    // but there doesn't seem to be an easy asynchronous solution.
-    // https://github.com/nodejs/node-v0.x-archive/issues/25330
-    if (fsx.existsSync(command + '.cmd')) {
-      command += '.cmd';
+  public static executeLifecycleCommand(
+    command: string,
+    options: ILifecycleCommandOptions
+  ): number {
+    const result: child_process.SpawnSyncReturns<Buffer> = Utilities._executeLifecycleCommandInternal(
+      command,
+      child_process.spawnSync,
+      options
+    );
+
+    if (options.handleOutput) {
+      Utilities._processResult(result);
     }
-
-    // This is needed since we specify shell=true below:
-    const escapedCommand: string = Utilities.escapeShellParameter(command);
-    const escapedArgs: string[] = args.map((x) => Utilities.escapeShellParameter(x));
-
-    return child_process.spawn(escapedCommand, escapedArgs, {
-      cwd: workingDirectory,
-      shell: true,
-      env: environmentVariables
-    });
+    return result.status;
   }
 
   /**
    * Executes the command using cmd if running on windows, or using sh if running on a non-windows OS.
    * @param command - the command to run on shell
-   * @param workingDirectory - working directory for running this command
-   * @param environmentVariables - environment variables for running this command
-   * @alpha
+   * @param options - options for how the command should be run
    */
-  public static executeShellCommand(
+  public static executeLifecycleCommandAsync(
     command: string,
-    workingDirectory: string,
-    environmentVariables?: { [key: string]: string },
-    captureOutput: boolean = false
-  ): child_process.SpawnSyncReturns<Buffer> {
-    let shellCommand: string = process.env.comspec || 'cmd';
-    let commandFlags: string = '/d /s /c';
-    let useShell: boolean = true;
-    if (process.platform !== 'win32') {
-      shellCommand = 'sh';
-      commandFlags = '-c';
-      useShell = false;
-    }
-
-    const result: child_process.SpawnSyncReturns<Buffer> = child_process.spawnSync(
-      shellCommand,
-      [commandFlags, command],
-      {
-        cwd: workingDirectory,
-        shell: useShell,
-        env: environmentVariables,
-        stdio: captureOutput ? ['pipe', 'pipe', 'pipe'] : [0, 1, 2]
-      });
-
-    Utilities._processResult(result);
-    return result;
-  }
-
-  /**
-   * Executes the command using cmd if running on windows, or using sh if running on a non-windows OS.
-   * @param command - the command to run on shell
-   * @param workingDirectory - working directory for running this command
-   * @param environmentVariables - environment variables for running this command
-   * @alpha
-   */
-  public static executeShellCommandAsync(
-    command: string,
-    workingDirectory: string,
-    environmentVariables?: { [key: string]: string },
-    captureOutput: boolean = false
+    options: ILifecycleCommandOptions
   ): child_process.ChildProcess {
-    let shellCommand: string = process.env.comspec || 'cmd';
-    let commandFlags: string = '/d /s /c';
-    let useShell: boolean = true;
-    if (process.platform !== 'win32') {
-      shellCommand = 'sh';
-      commandFlags = '-c';
-      useShell = false;
-    }
-
-    return child_process.spawn(
-      shellCommand,
-      [commandFlags, command],
-      {
-        cwd: workingDirectory,
-        shell: useShell,
-        env: environmentVariables,
-        stdio: captureOutput ? ['pipe', 'pipe', 'pipe'] : [0, 1, 2]
-      });
-  }
-
-  /**
-   * Returns the same thing as targetString.replace(searchValue, replaceValue), except that
-   * all matches are replaced, rather than just the first match.
-   * @param targetString  - The string to be modified
-   * @param searchValue   - The value to search for
-   * @param replaceValue  - The replacement text
-   */
-  public static getAllReplaced(targetString: string, searchValue: string, replaceValue: string): string {
-    return targetString.split(searchValue).join(replaceValue);
+    return Utilities._executeLifecycleCommandInternal(
+      command,
+      child_process.spawn,
+      options
+    );
   }
 
   /**
@@ -374,7 +435,243 @@ export default class Utilities {
    * Example: 'hello there' --> '"hello there"'
    */
   public static escapeShellParameter(parameter: string): string {
-    return '"' + parameter + '"';
+    return `"${parameter}"`;
+  }
+
+  /**
+   * Installs a package by name and version in the specified directory.
+   */
+  public static installPackageInDirectory(options: IInstallPackageInDirectoryOptions): void {
+    const directory: string = path.resolve(options.directory);
+    if (FileSystem.exists(directory)) {
+      console.log('Deleting old files from ' + directory);
+    }
+
+    FileSystem.ensureEmptyFolder(directory);
+
+    const npmPackageJson: IPackageJson = {
+      dependencies: {
+        [options.packageName]: options.version
+      },
+      description: 'Temporary file generated by the Rush tool',
+      name: options.tempPackageTitle,
+      private: true,
+      version: '0.0.0'
+    };
+    JsonFile.save(npmPackageJson, path.join(directory, FileConstants.PackageJson));
+
+    if (options.commonRushConfigFolder) {
+      Utilities.syncNpmrc(options.commonRushConfigFolder, directory);
+    }
+
+    console.log(os.EOL + 'Running "npm install" in ' + directory);
+
+    // NOTE: Here we use whatever version of NPM we happen to find in the PATH
+    Utilities.executeCommandWithRetry(
+      options.maxInstallAttempts,
+      'npm',
+      ['install'],
+      directory,
+      Utilities._createEnvironmentForRushCommand({}),
+      options.suppressOutput
+    );
+  }
+
+  public static withFinally<T>(options: { promise: Promise<T>, finally: () => void }): Promise<T> {
+    return options.promise.then<T>((result: T) => {
+      try {
+        options.finally();
+      } catch (error) {
+        return Promise.reject(error);
+      }
+      return result;
+    }).catch<T>((error: Error) => {
+      try {
+        options.finally();
+      } catch (innerError) {
+        return Promise.reject(innerError);
+      }
+      return Promise.reject(error);
+    });
+  }
+
+  /**
+   * As a workaround, syncNpmrc() copies the .npmrc file to the target folder, and also trims
+   * unusable lines from the .npmrc file.  If the source .npmrc file not exist, then syncNpmrc()
+   * will delete an .npmrc that is found in the target folder.
+   *
+   * Why are we trimming the .npmrc lines?  NPM allows environment variables to be specified in
+   * the .npmrc file to provide different authentication tokens for different registry.
+   * However, if the environment variable is undefined, it expands to an empty string, which
+   * produces a valid-looking mapping with an invalid URL that causes an error.  Instead,
+   * we'd prefer to skip that line and continue looking in other places such as the user's
+   * home directory.
+   *
+   * IMPORTANT: THIS CODE SHOULD BE KEPT UP TO DATE WITH _syncNpmrc() FROM scripts/install-run.ts
+   */
+  public static syncNpmrc(sourceNpmrcFolder: string, targetNpmrcFolder: string): void {
+    const sourceNpmrcPath: string = path.join(sourceNpmrcFolder, '.npmrc');
+    const targetNpmrcPath: string = path.join(targetNpmrcFolder, '.npmrc');
+    try {
+      if (FileSystem.exists(sourceNpmrcPath)) {
+        console.log(`Copying ${sourceNpmrcPath} --> ${targetNpmrcPath}`);
+        let npmrcFileLines: string[] = FileSystem.readFile(sourceNpmrcPath).split('\n');
+        npmrcFileLines = npmrcFileLines.map((line) => (line || '').trim());
+        const resultLines: string[] = [];
+        // Trim out lines that reference environment variables that aren't defined
+        for (const line of npmrcFileLines) {
+          // This finds environment variable tokens that look like "${VAR_NAME}"
+          const regex: RegExp = /\$\{([^\}]+)\}/g;
+          const environmentVariables: string[] | null = line.match(regex);
+          let lineShouldBeTrimmed: boolean = false;
+          if (environmentVariables) {
+            for (const token of environmentVariables) {
+              // Remove the leading "${" and the trailing "}" from the token
+              const environmentVariableName: string = token.substring(2, token.length - 1);
+              if (!process.env[environmentVariableName]) {
+                lineShouldBeTrimmed = true;
+                break;
+              }
+            }
+          }
+
+          if (lineShouldBeTrimmed) {
+            // Example output:
+            // "; MISSING ENVIRONMENT VARIABLE: //my-registry.com/npm/:_authToken=${MY_AUTH_TOKEN}"
+            resultLines.push('; MISSING ENVIRONMENT VARIABLE: ' + line);
+          } else {
+            resultLines.push(line);
+          }
+        }
+
+        FileSystem.writeFile(targetNpmrcPath, resultLines.join(os.EOL));
+      } else if (FileSystem.exists(targetNpmrcPath)) {
+        // If the source .npmrc doesn't exist and there is one in the target, delete the one in the target
+        console.log(`Deleting ${targetNpmrcPath}`);
+        FileSystem.deleteFile(targetNpmrcPath);
+      }
+    } catch (e) {
+      throw new Error(`Error syncing .npmrc file: ${e}`);
+    }
+  }
+
+  public static getRushConfigNotFoundError(): Error {
+    return new Error('Unable to find rush.json configuration file');
+  }
+
+  private static _executeLifecycleCommandInternal<TCommandResult>(
+    command: string,
+    spawnFunction: (command: String, args: string[], spawnOptions: child_process.SpawnOptions) => TCommandResult,
+    options: ILifecycleCommandOptions
+  ): TCommandResult {
+    let shellCommand: string = process.env.comspec || 'cmd';
+    let commandFlags: string = '/d /s /c';
+    let useShell: boolean = true;
+    if (process.platform !== 'win32') {
+      shellCommand = 'sh';
+      commandFlags = '-c';
+      useShell = false;
+    }
+
+    const environment: IEnvironment = Utilities._createEnvironmentForRushCommand(
+      {
+        initCwd: options.initCwd,
+        pathOptions: {
+          ...options.environmentPathOptions,
+          projectRoot: options.workingDirectory,
+          commonTempFolder: options.rushConfiguration ? options.rushConfiguration.commonTempFolder : undefined
+        }
+      }
+    );
+
+    return spawnFunction(
+      shellCommand,
+      [commandFlags, command],
+      {
+        cwd: options.workingDirectory,
+        shell: useShell,
+        env: environment,
+        stdio: options.handleOutput ? ['pipe', 'pipe', 'pipe'] : [0, 1, 2]
+      }
+    );
+  }
+
+  /**
+   * Returns a process.env environment suitable for executing lifecycle scripts.
+   * @param initialEnvironment - an existing environment to copy instead of process.env
+   */
+  private static _createEnvironmentForRushCommand(options: ICreateEnvironmentForRushCommandOptions): IEnvironment {
+    if (options.initialEnvironment === undefined) {
+      options.initialEnvironment = process.env;
+    }
+
+    const environment: IEnvironment = {};
+    for (const key of Object.getOwnPropertyNames(options.initialEnvironment)) {
+      const normalizedKey: string = os.platform() === 'win32' ? key.toUpperCase() : key;
+
+      // If Rush itself was invoked inside a lifecycle script, this may be set and would interfere
+      // with Rush's installations.  If we actually want it, we will set it explicitly below.
+      if (normalizedKey === 'INIT_CWD') {
+        continue;
+      }
+
+      // When NPM invokes a lifecycle event, it copies its entire configuration into environment
+      // variables.  Rush is supposed to be a deterministic controlled environment, so don't bring
+      // this along.
+      //
+      // NOTE: Longer term we should clean out the entire environment and use rush.json to bring
+      // back specific environment variables that the repo maintainer has determined to be safe.
+      if (normalizedKey.match(/^NPM_CONFIG_/)) {
+        continue;
+      }
+
+      // Use the uppercased environment variable name on Windows because environment variable names
+      // are case-insensitive on Windows
+      environment[normalizedKey] = options.initialEnvironment[key];
+    }
+
+    // When NPM invokes a lifecycle script, it sets an environment variable INIT_CWD that remembers
+    // the directory that NPM started in.  This allows naive scripts to change their current working directory
+    // and invoke NPM operations, while still be able to find a local .npmrc file.  Although Rush recommends
+    // for toolchain scripts to be professionally written (versus brittle stuff like
+    // "cd ./lib && npm run tsc && cd .."), we support INIT_CWD for compatibility.
+    //
+    // More about this feature: https://github.com/npm/npm/pull/12356
+    if (options.initCwd) {
+      environment['INIT_CWD'] = options.initCwd; // tslint:disable-line:no-string-literal
+    }
+
+    if (options.pathOptions) {
+      if (options.pathOptions.includeRepoBin && options.pathOptions.commonTempFolder) {
+        environment.PATH = Utilities._prependNodeModulesBinToPath(
+          environment.PATH,
+          options.pathOptions.commonTempFolder
+        );
+      }
+
+      if (options.pathOptions.includeProjectBin && options.pathOptions.projectRoot) {
+        environment.PATH = Utilities._prependNodeModulesBinToPath(
+          environment.PATH,
+          options.pathOptions.projectRoot
+        );
+      }
+    }
+
+    return environment;
+  }
+
+  /**
+   * Prepend the node_modules/.bin folder under the specified folder to the specified PATH variable. For example,
+   * if `rootDirectory` is "/foobar" and `existingPath` is "/bin", this function will return
+   * "/foobar/node_modules/.bin:/bin"
+   */
+  private static _prependNodeModulesBinToPath(existingPath: string | undefined, rootDirectory: string): string {
+    const binPath: string = path.resolve(rootDirectory, 'node_modules', '.bin');
+    if (existingPath) {
+      return `${binPath}${path.delimiter}${existingPath}`;
+    } else {
+      return binPath;
+    }
   }
 
   /**
@@ -383,19 +680,22 @@ export default class Utilities {
    */
   private static _executeCommandInternal(
     command: string, args: string[], workingDirectory: string,
-    stdio: (string|number)[],
-    environmentVariables: { [key: string]: string }): child_process.SpawnSyncReturns<Buffer> {
-
+    stdio: (string|number)[] | undefined,
+    environment?: IEnvironment,
+    keepEnvironment: boolean = false
+  ): child_process.SpawnSyncReturns<Buffer> {
     const options: child_process.SpawnSyncOptions = {
       cwd: workingDirectory,
       shell: true,
       stdio: stdio,
-      env: environmentVariables
+      env: keepEnvironment
+        ? environment
+        : Utilities._createEnvironmentForRushCommand({ initialEnvironment: environment })
     };
 
     // This is needed since we specify shell=true below.
-    // NOTE: On Windows if we escape "npm", the spawnSync() function runs something like this:
-    //   [ 'C:\\Windows\\system32\\cmd.exe', '/s', '/c', '""npm" "install""' ]
+    // NOTE: On Windows if we escape "NPM", the spawnSync() function runs something like this:
+    //   [ 'C:\\Windows\\system32\\cmd.exe', '/s', '/c', '""NPM" "install""' ]
     //
     // Due to a bug with Windows cmd.exe, the npm.cmd batch file's "%~dp0" variable will
     // return the current working directory instead of the batch file's directory.
@@ -406,7 +706,8 @@ export default class Utilities {
     // into node-core-library, but for now this hack will unblock people:
 
     // Only escape the command if it actually contains spaces:
-    const escapedCommand: string = command.indexOf(' ') < 0 ? command
+    const escapedCommand: string = command.indexOf(' ') < 0
+      ? command
       : Utilities.escapeShellParameter(command);
 
     const escapedArgs: string[] = args.map((x) => Utilities.escapeShellParameter(x));
@@ -414,13 +715,11 @@ export default class Utilities {
     let result: child_process.SpawnSyncReturns<Buffer> = child_process.spawnSync(escapedCommand,
       escapedArgs, options);
 
-    /* tslint:disable:no-any */
-    if (result.error && (result.error as any).errno === 'ENOENT') {
+    if (result.error && (result.error as any).errno === 'ENOENT') { // tslint:disable-line:no-any
       // This is a workaround for GitHub issue #25330
       // https://github.com/nodejs/node-v0.x-archive/issues/25330
       result = child_process.spawnSync(command + '.cmd', args, options);
     }
-    /* tslint:enable:no-any */
 
     Utilities._processResult(result);
     return result;
