@@ -1,6 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-/// <reference path="git-repo-info/git-repo-info.d.ts" />
-/// <reference path="jju/jju.d.ts" />
-/// <reference path="rimraf/rimraf.d.ts" />
+/// <reference path="builtins/builtins.d.ts" />
+/// <reference path="glob-escape/glob-escape.d.ts" />
+/// <reference path="jest/jest.d.ts" />
+/// <reference path="mocks/child_process.d.ts" />
+/// <reference path="npm-package-arg/npm-package-arg.d.ts" />
+/// <reference path="read-package-tree/read-package-tree.d.ts" />
+/// <reference path="strict-uri-encode/strict-uri-encode.d.ts" />
+/// <reference path="wordwrap/wordwrap.d.ts" />
+/// <reference path="yarnpkg-lockfile/yarnpkg-lockfile.d.ts" />
