@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { GulpTask } from '@microsoft/gulp-core-build';
+import { GulpTask, IBuildConfig } from '@microsoft/gulp-core-build';
 import * as Gulp from 'gulp';
 import * as gulpIstanbul from 'gulp-istanbul';
 
@@ -16,6 +16,13 @@ export class InstrumentTask extends GulpTask<IInstrumentTaskConfig> {
       {
         coverageMatch: ['lib/**/*.js', '!lib/**/*.test.js']
       }
+    );
+  }
+
+  public isEnabled(buildConfig: IBuildConfig): boolean {
+    return (
+      super.isEnabled(buildConfig) &&
+      !buildConfig.jestEnabled
     );
   }
 
