@@ -2,12 +2,11 @@
 // See LICENSE in the project root for license information.
 
 import * as path from 'path';
-import * as fs from 'fs';
 import { homedir } from 'os';
 
-const encoding: string = 'utf8';
+import { FileSystem } from '@microsoft/node-core-library';
 
-export default class CertificateStore {
+export class CertificateStore {
   private static _instance: CertificateStore;
 
   public static get instance(): CertificateStore {
@@ -33,8 +32,8 @@ export default class CertificateStore {
 
   public get certificateData(): string {
     if (!this._certificateData) {
-      if (fs.existsSync(this._certificatePath)) {
-        this._certificateData = fs.readFileSync(this._certificatePath, encoding);
+      if (FileSystem.exists(this._certificatePath)) {
+        this._certificateData = FileSystem.readFile(this._certificatePath);
       } else {
         return undefined;
       }
@@ -45,9 +44,9 @@ export default class CertificateStore {
 
   public set certificateData(certificate: string) {
     if (certificate) {
-      fs.writeFileSync(this._certificatePath, certificate);
-    } else if (fs.existsSync(this._certificatePath)) {
-      fs.unlinkSync(this._certificatePath);
+      FileSystem.writeFile(this._certificatePath, certificate);
+    } else if (FileSystem.exists(this._certificatePath)) {
+      FileSystem.deleteFile(this._certificatePath);
     }
 
     this._certificateData = certificate;
@@ -55,8 +54,8 @@ export default class CertificateStore {
 
   public get keyData(): string {
     if (!this._keyData) {
-      if (fs.existsSync(this._keyPath)) {
-        this._keyData = fs.readFileSync(this._keyPath, encoding);
+      if (FileSystem.exists(this._keyPath)) {
+        this._keyData = FileSystem.readFile(this._keyPath);
       } else {
         return undefined;
       }
@@ -67,9 +66,9 @@ export default class CertificateStore {
 
   public set keyData(key: string) {
     if (key) {
-      fs.writeFileSync(this._keyPath, key, { encoding });
-    } else if (fs.existsSync(this._keyPath)) {
-      fs.unlinkSync(this._keyPath);
+      FileSystem.writeFile(this._keyPath, key);
+    } else if (FileSystem.exists(this._keyPath)) {
+      FileSystem.deleteFile(this._keyPath);
     }
 
     this._keyData = key;
@@ -78,14 +77,12 @@ export default class CertificateStore {
   private _initialize(): void {
     const unresolvedUserFolder: string = homedir();
     this._userProfilePath = path.resolve(unresolvedUserFolder);
-    if (!fs.existsSync(this._userProfilePath)) {
+    if (!FileSystem.exists(this._userProfilePath)) {
       throw new Error('Unable to determine the current user\'s home directory');
     }
 
     this._gcbServeDataPath = path.join(this._userProfilePath, '.gcb-serve-data');
-    if (!fs.existsSync(this._gcbServeDataPath)) {
-      fs.mkdirSync(this._gcbServeDataPath);
-    }
+    FileSystem.ensureFolder(this._gcbServeDataPath);
 
     this._certificatePath = path.join(this._gcbServeDataPath, 'gcb-serve.cer');
     this._keyPath = path.join(this._gcbServeDataPath, 'gcb-serve.key');
