@@ -3,7 +3,10 @@
 
 import { GulpTask } from '@microsoft/gulp-core-build';
 import * as Gulp from 'gulp';
-import { ICertificate } from './certificates';
+import {
+  ICertificate,
+  ensureCertificate
+} from './certificates';
 
 /**
  * This task generates and trusts a development certificate. The certificate is self-signed
@@ -12,13 +15,12 @@ import { ICertificate } from './certificates';
  *  trusted as a root cert in the keychain. On other platforms, the certificate is generated
  *  and signed, but the user must trust it manually.
  */
-export default class TrustCertTask extends GulpTask<void> {
+export class TrustCertTask extends GulpTask<void> {
   constructor() {
     super('trust-cert');
   }
 
   public executeTask(gulp: typeof Gulp, completeCallback: (error?: string) => void): void {
-    const { ensureCertificate } = require('./certificates'); // tslint:disable-line
     const certificate: ICertificate = ensureCertificate(true, this);
 
     if (certificate.pemCertificate && certificate.pemKey) {
