@@ -12,11 +12,13 @@ export interface ISudoSyncResult {
 }
 
 export function runSudoSync(params: string[]): ISudoSyncResult {
-  const sudoResult: child_process.ChildProcess = sudo(params,
-                                                      {
-                                                        cachePassword: false,
-                                                        prompt: 'Enter your password: '
-                                                      });
+  const sudoResult: child_process.ChildProcess = sudo(
+    params,
+    {
+      cachePassword: false,
+      prompt: 'Enter your password: '
+    }
+  );
 
   const stderr: string[] = [];
   sudoResult.stderr.on('data', (data: Buffer) => {
