@@ -1,22 +1,28 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { TypeScriptTask } from './TypeScriptTask';
-import { TSLintTask } from './TSLintTask';
-import { TextTask } from './TextTask';
-import { RemoveTripleSlashReferenceTask } from './RemoveTripleSlashReferenceTask';
-import { IExecutable, parallel, serial } from '@microsoft/gulp-core-build';
+import {
+  TscCmdTask,
+  ITscCmdTaskConfig
+} from './TscCmdTask';
+import {
+  TslintCmdTask,
+  ITslintCmdTaskConfig
+} from './TslintCmdTask';
 import { ApiExtractorTask } from './ApiExtractorTask';
 
-export * from './TypeScriptConfiguration';
-export { TypeScriptTask } from './TypeScriptTask';
-export { ApiExtractorTask } from './ApiExtractorTask';
+export {
+  TscCmdTask,
+  ITscCmdTaskConfig,
+  TslintCmdTask,
+  ITslintCmdTaskConfig
+};
 
+/** @public */
+export const tscCmd: TscCmdTask = new TscCmdTask();
+
+/** @public */
+export const tslintCmd: TslintCmdTask = new TslintCmdTask();
+
+/** @public */
 export const apiExtractor: ApiExtractorTask = new ApiExtractorTask();
-export const typescript: TypeScriptTask = new TypeScriptTask();
-export const tslint: TSLintTask = new TSLintTask();
-export const text: TextTask = new TextTask();
-export const removeTripleSlash: RemoveTripleSlashReferenceTask = new RemoveTripleSlashReferenceTask();
-
-// tslint:disable:export-name
-export default parallel(tslint, serial(typescript, removeTripleSlash)) as IExecutable;
