@@ -1,5 +1,5 @@
 'use strict';
 
 let build = require('@microsoft/node-library-build');
-
+build.mocha.enabled = false;
 build.initialize(require('gulp'));
