@@ -2,23 +2,22 @@
 // See LICENSE in the project root for license information.
 
 import { args } from './State';
+import { getConfig } from './index';
 
 const ENVIRONMENT_VARIABLE_PREFIX: string = 'GCB_';
-
-let _defaultValues: Object = {};
-
-export function setConfigDefaults(defaultValues: Object): void {
-  _defaultValues = defaultValues;
-}
 
 export function getConfigValue(name: string, defaultValue?: string | boolean): string | boolean {
 
   // Try to get config value from environment variable.
   const envVariable: string = ENVIRONMENT_VARIABLE_PREFIX + name.toUpperCase();
-  const envValue: string = process.env[envVariable];
+  const envValue: string | undefined = process.env[envVariable];
   const argsValue: string | boolean = args[name.toLowerCase()];
 
-  return _firstDefinedValue(argsValue, envValue, defaultValue, _defaultValues[name]);
+  // getConfig can be undefined during the first few calls to this function because the build config is initialized
+  // before the getConfig function is defined. In those cases, a defaultValue is provided.
+  const configValue: string | boolean = ((getConfig ? getConfig() : {}) || {})[name];
+
+  return _firstDefinedValue(argsValue, envValue, defaultValue, configValue);
 }
 
 export function getFlagValue(name: string, defaultValue?: boolean): boolean {
