@@ -1,11 +1,14 @@
 'use strict';
 
 let build = require('@microsoft/gulp-core-build');
-let apiExtractor = require('@microsoft/gulp-core-build-typescript').apiExtractor;
-let typescript = require('@microsoft/gulp-core-build-typescript').typescript;
+let { tscCmd, tslintCmd, apiExtractor } = require('@microsoft/gulp-core-build-typescript')
 let mocha = require('@microsoft/gulp-core-build-mocha');
 
-build.task('default', build.serial(typescript, mocha, apiExtractor));
+build.setConfig({
+  shouldWarningsFailBuild: build.getConfig().production
+});
+
+build.task('default', build.serial(build.parallel(tscCmd, tslintCmd), apiExtractor, mocha));
 
 build.initialize(require('gulp'));
 
