@@ -4,15 +4,15 @@
 import {
   CopyTask,
   copyStaticAssets,
+  jest,
   task,
   watch,
   serial,
   parallel,
   IExecutable,
-  setConfig,
-  IBuildConfig
+  setConfig
 } from '@microsoft/gulp-core-build';
-import { typescript, tslint, apiExtractor } from '@microsoft/gulp-core-build-typescript';
+import { tscCmd, tslintCmd, apiExtractor } from '@microsoft/gulp-core-build-typescript';
 import { instrument, mocha } from '@microsoft/gulp-core-build-mocha';
 
 export * from '@microsoft/gulp-core-build';
@@ -20,25 +20,44 @@ export * from '@microsoft/gulp-core-build-typescript';
 export * from '@microsoft/gulp-core-build-mocha';
 
 // pre copy and post copy allows you to specify a map of dest: [sources] to copy from one place to another.
+/**
+ * @public
+ */
 export const preCopy: CopyTask = new CopyTask();
 preCopy.name = 'pre-copy';
 
+/**
+ * @public
+ */
 export const postCopy: CopyTask = new CopyTask();
 postCopy.name = 'post-copy';
 
-const PRODUCTION = process.argv.indexOf('--production') !== -1 || process.argv.indexOf('--ship') !== -1;
+const PRODUCTION: boolean = process.argv.indexOf('--production') !== -1 || process.argv.indexOf('--ship') !== -1;
 setConfig({
   production: PRODUCTION,
   shouldWarningsFailBuild: PRODUCTION
 });
 
-tslint.mergeConfig({
-  displayAsWarning: true
-});
+const buildSubtask: IExecutable = serial(
+  preCopy,
+  parallel(tslintCmd, tscCmd, copyStaticAssets),
+  apiExtractor,
+  postCopy
+);
 
-const buildSubtask: IExecutable = serial(preCopy, parallel(tslint, typescript, copyStaticAssets), apiExtractor, postCopy);
+/**
+ * @public
+ */
 export const buildTasks: IExecutable = task('build', buildSubtask);
-export const testTasks: IExecutable = task('test', serial(buildSubtask, mocha));
-export const defaultTasks: IExecutable = task('default', serial(buildSubtask, instrument, mocha));
+
+/**
+ * @public
+ */
+export const testTasks: IExecutable = task('test', serial(buildSubtask, mocha, jest));
+
+/**
+ * @public
+ */
+export const defaultTasks: IExecutable = task('default', serial(buildSubtask, instrument, mocha, jest));
 
 task('watch', watch('src/**.ts', testTasks));
