@@ -1,10 +1,5 @@
 'use strict';
 
-let build = require('@microsoft/gulp-core-build');
-let typescript = require('@microsoft/gulp-core-build-typescript').typescript;
-let apiExtractor = require('@microsoft/gulp-core-build-typescript').apiExtractor;
-
-build.task('default', build.serial(typescript, apiExtractor));
-
+let build = require('@microsoft/node-library-build');
+build.mocha.enabled = false;
 build.initialize(require('gulp'));
-
