@@ -5,18 +5,17 @@ import {
   CopyTask,
   GenerateShrinkwrapTask,
   IExecutable,
+  jest,
   ValidateShrinkwrapTask,
   parallel,
   serial,
   task,
   watch,
   setConfig,
-  getConfig,
-  IBuildConfig
+  getConfig
 } from '@microsoft/gulp-core-build';
-import { apiExtractor, typescript, tslint, text } from '@microsoft/gulp-core-build-typescript';
+import { apiExtractor, tscCmd, tslintCmd } from '@microsoft/gulp-core-build-typescript';
 import { sass } from '@microsoft/gulp-core-build-sass';
-import { karma } from '@microsoft/gulp-core-build-karma';
 import { webpack } from '@microsoft/gulp-core-build-webpack';
 import { serve, reload } from '@microsoft/gulp-core-build-serve';
 import { PostProcessSourceMaps } from './PostProcessSourceMaps';
@@ -24,14 +23,19 @@ import { PostProcessSourceMaps } from './PostProcessSourceMaps';
 export * from '@microsoft/gulp-core-build';
 export * from '@microsoft/gulp-core-build-typescript';
 export * from '@microsoft/gulp-core-build-sass';
-export * from '@microsoft/gulp-core-build-karma';
 export * from '@microsoft/gulp-core-build-webpack';
 export * from '@microsoft/gulp-core-build-serve';
 
-// pre copy and post copy allows you to specify a map of dest: [sources] to copy from one place to another.
+// Pre copy and post copy allows you to specify a map of dest: [sources] to copy from one place to another.
+/**
+ * @public
+ */
 export const preCopy: CopyTask = new CopyTask();
 preCopy.name = 'pre-copy';
 
+/**
+ * @public
+ */
 export const postCopy: CopyTask = new CopyTask();
 postCopy.name = 'post-copy';
 
@@ -40,24 +44,50 @@ const sourceMatch: string[] = [
   '!src/**/*.scss.ts'
 ];
 
-const PRODUCTION = !!getConfig().args['production'] || !!getConfig().args['ship'];
+// tslint:disable-next-line:no-string-literal
+const PRODUCTION: boolean = !!getConfig().args['production'] || !!getConfig().args['ship'];
 setConfig({
   production: PRODUCTION,
   shouldWarningsFailBuild: PRODUCTION
 });
 
-tslint.mergeConfig({
-  displayAsWarning: true
-});
-
 // Define default task groups.
-export const compileTsTasks: IExecutable = parallel(typescript, text, apiExtractor);
-export const buildTasks: IExecutable = task('build', serial(preCopy, sass, parallel(tslint, compileTsTasks), postCopy));
+/**
+ * @public
+ */
+export const buildTasks: IExecutable = task(
+  'build',
+  serial(preCopy, sass, parallel(tslintCmd, tscCmd), apiExtractor, postCopy)
+);
+
+/**
+ * @public
+ */
 export const bundleTasks: IExecutable = task('bundle', serial(buildTasks, webpack));
-export const testTasks: IExecutable = task('test', serial(buildTasks, karma));
-export const defaultTasks: IExecutable = serial(bundleTasks, karma);
+
+/**
+ * @public
+ */
+export const testTasks: IExecutable = task('test', serial(buildTasks, jest));
+
+/**
+ * @public
+ */
+export const defaultTasks: IExecutable = serial(bundleTasks, jest);
+
+/**
+ * @public
+ */
 export const postProcessSourceMapsTask: PostProcessSourceMaps = new PostProcessSourceMaps();
+
+/**
+ * @public
+ */
 export const validateShrinkwrapTask: ValidateShrinkwrapTask = new ValidateShrinkwrapTask();
+
+/**
+ * @public
+ */
 export const generateShrinkwrapTask: GenerateShrinkwrapTask = new GenerateShrinkwrapTask();
 
 task('validate-shrinkwrap', validateShrinkwrapTask);
@@ -67,11 +97,12 @@ task('test-watch', watch(sourceMatch, testTasks));
 // For watch scenarios like serve, make sure to exclude generated files from src (like *.scss.ts.)
 task('serve',
   serial(
-  serve,
-  watch(
-    sourceMatch, serial(preCopy, sass, compileTsTasks,
-      postCopy, webpack, postProcessSourceMapsTask, reload)
-  ))
+    serve,
+    watch(
+      sourceMatch,
+      serial(preCopy, sass, tscCmd, postCopy, webpack, postProcessSourceMapsTask, reload)
+    )
+  )
 );
 
 task('default', defaultTasks);
