@@ -1,9 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-/// <reference types='mocha' />
-
-import { assert } from 'chai';
 import * as path from 'path';
 import { JsonFile } from '../JsonFile';
 import { JsonSchema, IJsonSchemaErrorInfo } from '../JsonSchema';
@@ -18,14 +15,22 @@ describe('JsonSchema', () => {
 
   const schema: JsonSchema = JsonSchema.fromFile(schemaPath);
 
-  it('loadAndValidate successfully validates a JSON file', (done: MochaDone) => {
+  test('loadAndValidate successfully validates a JSON file', () => {
     const jsonPath: string = path.resolve(path.join(__dirname, './test-data/test.json'));
     const jsonObject: Object = JsonFile.loadAndValidate(jsonPath, schema);
-    assert.isObject(jsonObject);
-    done();
+    expect(jsonObject).toMatchObject(
+      {
+        'exampleString': 'This is a string',
+        'exampleArray': [
+          'apple',
+          'banana',
+          'coconut'
+        ]
+      }
+    );
   });
 
-  it('validateObjectWithCallback successfully reports a compound validation error', (done: MochaDone) => {
+  test('validateObjectWithCallback successfully reports a compound validation error', () => {
     const jsonPath2: string = path.resolve(path.join(__dirname, './test-data/test2.json'));
     const jsonObject2: Object = JsonFile.load(jsonPath2);
 
@@ -41,12 +46,9 @@ Error: #/exampleOneOf (Description for type2)
 
     schema.validateObjectWithCallback(jsonObject2, (errorInfo: IJsonSchemaErrorInfo) => {
       ++errorCount;
-      console.log(errorInfo.details);
-      assert.equal(normalize(errorInfo.details), normalize(expectedError),
-        'Error #' + errorCount.toString());
+      expect(normalize(errorInfo.details)).toEqual(normalize(expectedError));
     });
 
-    assert.equal(errorCount, 1);
-    done();
+    expect(errorCount).toEqual(1);
   });
 });
