@@ -5,8 +5,12 @@ import { getPackageDeps, parseGitLsTree } from '../getPackageDeps';
 import { IPackageDeps } from '../IPackageDeps';
 import { expect, assert } from 'chai';
 import * as path from 'path';
-import * as fs from 'fs';
 import { execSync } from 'child_process';
+
+import {
+  FileSystem,
+  FileConstants
+} from '@microsoft/node-core-library';
 
 const SOURCE_PATH: string = path.join(__dirname).replace(
   path.join('lib', 'test'),
@@ -64,12 +68,12 @@ describe('parseGitLsTree', () => {
 
 describe('getPackageDeps', () => {
 
-  it('can parse commited file', (done) => {
+  it('can parse committed file', (done) => {
     const results: IPackageDeps = getPackageDeps(TEST_PROJECT_PATH);
     try {
       const expectedFiles: { [key: string]: string } = {
         'file1.txt': 'c7b2f707ac99ca522f965210a7b6b0b109863f34',
-        'package.json': '33703d582243a41bdebff8ee7dd046a01fc054b9'
+        [FileConstants.PackageJson]: '33703d582243a41bdebff8ee7dd046a01fc054b9'
       };
       const filePaths: string[] = Object.keys(results.files).sort();
 
@@ -87,7 +91,7 @@ describe('getPackageDeps', () => {
     try {
       const expectedFiles: { [key: string]: string } = {
         'src/file 1.txt': 'c7b2f707ac99ca522f965210a7b6b0b109863f34',
-        'package.json': '33703d582243a41bdebff8ee7dd046a01fc054b9'
+        [FileConstants.PackageJson]: '33703d582243a41bdebff8ee7dd046a01fc054b9'
       };
       const filePaths: string[] = Object.keys(results.files).sort();
 
@@ -100,13 +104,13 @@ describe('getPackageDeps', () => {
     done();
   });
 
-  it('can can handle adding one file', (done) => { // tslint:disable-line
+  it('can handle adding one file', (done) => { // tslint:disable-line
     const tempFilePath: string = path.join(TEST_PROJECT_PATH, 'a.txt');
 
-    fs.writeFileSync(tempFilePath, 'a');
+    FileSystem.writeFile(tempFilePath, 'a');
 
     function _done(e?: Error): void {
-      fs.unlinkSync(tempFilePath);
+      FileSystem.deleteFile(tempFilePath);
       done(e);
     }
 
@@ -115,7 +119,7 @@ describe('getPackageDeps', () => {
       const expectedFiles: { [key: string]: string } = {
         'a.txt': '2e65efe2a145dda7ee51d1741299f848e5bf752e',
         'file1.txt': 'c7b2f707ac99ca522f965210a7b6b0b109863f34',
-        'package.json': '33703d582243a41bdebff8ee7dd046a01fc054b9'
+        [FileConstants.PackageJson]: '33703d582243a41bdebff8ee7dd046a01fc054b9'
       };
       const filePaths: string[] = Object.keys(results.files).sort();
 
@@ -132,16 +136,16 @@ describe('getPackageDeps', () => {
 
   });
 
-  it('can can handle adding two files', (done) => { // tslint:disable-line
+  it('can handle adding two files', (done) => { // tslint:disable-line
     const tempFilePath1: string = path.join(TEST_PROJECT_PATH, 'a.txt');
     const tempFilePath2: string = path.join(TEST_PROJECT_PATH, 'b.txt');
 
-    fs.writeFileSync(tempFilePath1, 'a');
-    fs.writeFileSync(tempFilePath2, 'a');
+    FileSystem.writeFile(tempFilePath1, 'a');
+    FileSystem.writeFile(tempFilePath2, 'a');
 
     function _done(e?: Error): void {
-      fs.unlinkSync(tempFilePath1);
-      fs.unlinkSync(tempFilePath2);
+      FileSystem.deleteFile(tempFilePath1);
+      FileSystem.deleteFile(tempFilePath2);
       done(e);
     }
 
@@ -151,7 +155,7 @@ describe('getPackageDeps', () => {
         'a.txt': '2e65efe2a145dda7ee51d1741299f848e5bf752e',
         'b.txt': '2e65efe2a145dda7ee51d1741299f848e5bf752e',
         'file1.txt': 'c7b2f707ac99ca522f965210a7b6b0b109863f34',
-        'package.json': '33703d582243a41bdebff8ee7dd046a01fc054b9'
+        [FileConstants.PackageJson]: '33703d582243a41bdebff8ee7dd046a01fc054b9'
       };
       const filePaths: string[] = Object.keys(results.files).sort();
 
@@ -167,20 +171,20 @@ describe('getPackageDeps', () => {
     _done();
   });
 
-  it('can can handle removing one file', (done) => {
+  it('can handle removing one file', (done) => {
     const testFilePath: string = path.join(TEST_PROJECT_PATH, 'file1.txt');
 
-    fs.unlinkSync(testFilePath);
+    FileSystem.deleteFile(testFilePath);
 
     function _done(e?: Error): void {
-      execSync(`git checkout ${ testFilePath }`);
+      execSync(`git checkout ${ testFilePath }`, { stdio: 'ignore' });
       done(e);
     }
 
     const results: IPackageDeps = getPackageDeps(TEST_PROJECT_PATH);
     try {
       const expectedFiles: { [key: string]: string } = {
-        'package.json': '33703d582243a41bdebff8ee7dd046a01fc054b9'
+        [FileConstants.PackageJson]: '33703d582243a41bdebff8ee7dd046a01fc054b9'
       };
       const filePaths: string[] = Object.keys(results.files).sort();
 
@@ -195,13 +199,13 @@ describe('getPackageDeps', () => {
     _done();
   });
 
-  it('can can handle changing one file', (done) => {
+  it('can handle changing one file', (done) => {
     const testFilePath: string = path.join(TEST_PROJECT_PATH, 'file1.txt');
 
-    fs.writeFileSync(testFilePath, 'abc');
+    FileSystem.writeFile(testFilePath, 'abc');
 
     function _done(e?: Error): void {
-      execSync(`git checkout ${testFilePath}`);
+      execSync(`git checkout ${testFilePath}`, { stdio: 'ignore' });
       done(e);
     }
 
@@ -209,7 +213,7 @@ describe('getPackageDeps', () => {
     try {
       const expectedFiles: { [key: string]: string } = {
         'file1.txt': 'f2ba8f84ab5c1bce84a7b441cb1959cfc7093b7f',
-        'package.json': '33703d582243a41bdebff8ee7dd046a01fc054b9'
+        [FileConstants.PackageJson]: '33703d582243a41bdebff8ee7dd046a01fc054b9'
       };
       const filePaths: string[] = Object.keys(results.files).sort();
 
@@ -228,7 +232,7 @@ describe('getPackageDeps', () => {
     const results: IPackageDeps = getPackageDeps(TEST_PROJECT_PATH, ['file1.txt']);
     try {
       const expectedFiles: { [key: string]: string } = {
-        'package.json': '33703d582243a41bdebff8ee7dd046a01fc054b9'
+        [FileConstants.PackageJson]: '33703d582243a41bdebff8ee7dd046a01fc054b9'
       };
       const filePaths: string[] = Object.keys(results.files).sort();
 
@@ -244,10 +248,10 @@ describe('getPackageDeps', () => {
   it('can exclude an added file', (done) => {
     const tempFilePath: string = path.join(TEST_PROJECT_PATH, 'a.txt');
 
-    fs.writeFileSync(tempFilePath, 'a');
+    FileSystem.writeFile(tempFilePath, 'a');
 
     function _done(e?: Error): void {
-      fs.unlinkSync(tempFilePath);
+      FileSystem.deleteFile(tempFilePath);
       done(e);
     }
 
@@ -255,7 +259,7 @@ describe('getPackageDeps', () => {
     try {
       const expectedFiles: { [key: string]: string } = {
         'file1.txt': 'c7b2f707ac99ca522f965210a7b6b0b109863f34',
-        'package.json': '33703d582243a41bdebff8ee7dd046a01fc054b9'
+        [FileConstants.PackageJson]: '33703d582243a41bdebff8ee7dd046a01fc054b9'
       };
       const filePaths: string[] = Object.keys(results.files).sort();
 
