@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-export { default as Interleaver } from './Interleaver';
-export { ITaskWriter as ITaskWriter } from './Interleaver';
-export { default as StreamCollator } from './next/StreamCollator';
-export { default as DualTaskStream } from './next/DualTaskStream';
-export { default as PersistentStream } from './next/PersistentStream';
+export {
+  Interleaver,
+  ITaskWriter
+} from './Interleaver';
