@@ -4,10 +4,9 @@
 /// <reference types="mocha" />
 
 import { assert } from 'chai';
-import * as colors from 'colors';
 import * as os from 'os';
 
-import Interleaver, { ITaskWriter } from '../Interleaver';
+import { Interleaver, ITaskWriter } from '../Interleaver';
 
 class StringStream {
   private _buffer: string[] = [];
@@ -36,15 +35,6 @@ describe('Interleaver tests', () => {
   });
 
   describe('Testing register and close', () => {
-    it('cannot be directly instantiated', (done: MochaDone) => {
-      assert.throws(() => {
-        /* tslint:disable:no-unused-variable */
-        const a: Interleaver = new Interleaver();
-        /* tslint:enable:no-unused-variable */
-      });
-      done();
-    });
-
     it('can register a task', (done: MochaDone) => {
       const helloWorldWriter: ITaskWriter = Interleaver.registerTask('Hello World');
       assert.isObject(helloWorldWriter);
@@ -86,26 +76,12 @@ describe('Interleaver tests', () => {
       done();
     });
 
-    it('should redirect warnings to stdout in yellow', (done: MochaDone) => {
-      const taskA: ITaskWriter = Interleaver.registerTask('A');
-      const warning: string = 'Warning - This is a warning';
-
-      taskA.writeError(warning);
-      assert.equal(stdout.read(), colors.yellow(warning));
-
-      taskA.close();
-
-      assert.equal(taskA.getStdOutput(), warning);
-      assert.equal(taskA.getStdError(), '');
-      done();
-    });
-
-    it('should write errors in red', (done: MochaDone) => {
+    it('should write errors to stderr', (done: MochaDone) => {
       const taskA: ITaskWriter = Interleaver.registerTask('A');
       const error: string = 'Critical error';
 
       taskA.writeError(error);
-      assert.equal(stdout.read(), colors.red(error));
+      assert.equal(stdout.read(), error);
 
       taskA.close();
 
