@@ -1,7 +1,5 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
- * See LICENSE in the project root for license information.
- */
+// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+// See LICENSE in the project root for license information.
 
 import { EOL } from 'os';
 import {
@@ -9,7 +7,17 @@ import {
   escapeRegExp
 } from 'lodash';
 import * as Webpack from 'webpack';
-import * as ITapable from 'tapable';
+import * as Tapable from 'tapable';
+
+import {
+  IV3Compilation,
+  IV3Module,
+  IV3Chunk
+} from './V3Interfaces';
+import {
+  IInternalOptions,
+  getSetPublicPathCode
+} from './codeGenerator';
 
 /**
  * The base options for setting the webpack public path at runtime.
@@ -51,6 +59,17 @@ export interface ISetWebpackPublicPathOptions {
    * See the README for more information.
    */
   getPostProcessScript?: (varName: string) => string;
+
+  /**
+   * If true, find the last script matching the regexVariable (if it is set). If false, find the first matching script.
+   * This can be useful if there are multiple scripts loaded in the DOM that match the regexVariable.
+   */
+  preferLastFoundScript?: boolean;
+
+  /**
+   * If true, always include the public path-setting code. Don't try to detect if any chunks or assets are present.
+   */
+  skipDetection?: boolean;
 }
 
 /**
@@ -77,31 +96,22 @@ export interface ISetWebpackPublicPathPluginOptions extends ISetWebpackPublicPat
   };
 }
 
-import {
-  IInternalOptions,
-  getSetPublicPathCode
-} from './codeGenerator';
-
-interface IAsset {
-}
-
-interface IChunk {
-  chunks: IChunk[];
-  name: string;
-  renderedHash: string;
-  forEachModule(iterator: (module: IModule) => void): void;
-}
-
-interface IModule {
-  assets: IAsset[];
-}
-
-interface IMainTemplate extends ITapable {
+interface IV4MainTemplate extends Webpack.compilation.MainTemplate {
+  hooks: {
+    startup: Tapable.SyncHook<string, Webpack.compilation.Chunk, string>;
+  };
   requireFn: string;
 }
 
-interface ICompilation {
-  mainTemplate: IMainTemplate;
+interface IV4Chunk extends Webpack.compilation.Chunk {
+  forEachModule(callback: (module: Webpack.compilation.Module) => void): void;
+}
+
+interface IStartupCodeOptions {
+  source: string;
+  chunk: IV3Chunk | Webpack.compilation.Chunk;
+  hash: string;
+  requireFn: string;
 }
 
 /**
@@ -118,44 +128,94 @@ export class SetPublicPathPlugin implements Webpack.Plugin {
   }
 
   public apply(compiler: Webpack.Compiler): void {
-    compiler.plugin('compilation', (compilation: ICompilation, params: Object): void => {
-      compilation.mainTemplate.plugin('startup', (source: string, chunk: IChunk, hash: string) => {
-        let assetOrChunkFound: boolean = chunk.chunks.length > 0;
-        if (!assetOrChunkFound) {
-          chunk.forEachModule((innerModule: IModule) => {
-            if (innerModule.assets && Object.keys(innerModule.assets).length > 0) {
-              assetOrChunkFound = true;
+    const isWebpack4: boolean = !!compiler.hooks;
+
+    if (isWebpack4) {
+      compiler.hooks.compilation.tap('set-webpack-public-path', (compilation: Webpack.compilation.Compilation) => {
+        const v4MainTemplate: IV4MainTemplate = compilation.mainTemplate as IV4MainTemplate;
+        v4MainTemplate.hooks.startup.tap(
+          'set-webpack-public-path',
+          (source: string, chunk: IV4Chunk, hash: string) => {
+            let assetOrChunkFound: boolean = !!this.options.skipDetection;
+
+            if (!assetOrChunkFound) {
+              for (const chunkGroup of chunk.groupsIterable) {
+                const children: Webpack.compilation.Chunk[] = chunkGroup.getChildren();
+                assetOrChunkFound = assetOrChunkFound || (children.length > 0);
+              }
             }
-          });
-        }
 
-        if (assetOrChunkFound) {
-          const moduleOptions: IInternalOptions = cloneDeep(this.options);
+            if (!assetOrChunkFound) {
+              for (const innerModule of chunk.modulesIterable) {
+                if (innerModule.buildInfo.assets && Object.keys(innerModule.buildInfo.assets).length > 0) {
+                  assetOrChunkFound = true;
+                }
+              }
+            }
 
-          // If this module has ownership over any chunks or assets, inject the public path code
-          moduleOptions.webpackPublicPathVariable = `${compilation.mainTemplate.requireFn}.p`;
-          moduleOptions.linePrefix = '  ';
-
-          if (this.options.scriptName) {
-            moduleOptions.regexName = this.options.scriptName.name;
-            if (this.options.scriptName.isTokenized) {
-              moduleOptions.regexName = moduleOptions.regexName.replace(/\[name\]/g, escapeRegExp(chunk.name))
-                                                               .replace(/\[hash\]/g, chunk.renderedHash);
+            if (assetOrChunkFound) {
+              return this._getStartupCode({
+                source,
+                chunk,
+                hash,
+                requireFn: v4MainTemplate.requireFn
+              });
+            } else {
+              return source;
             }
           }
-
-          return [
-            '// Set the webpack public path',
-            '(function () {',
-              getSetPublicPathCode(moduleOptions, console.error),
-            '})();',
-            '',
-            source
-          ].join(EOL);
-        } else {
-          return source;
-        }
+        );
       });
-    });
+    } else {
+      compiler.plugin('compilation', (compilation: IV3Compilation, params: Object): void => {
+        compilation.mainTemplate.plugin('startup', (source: string, chunk: IV3Chunk, hash: string) => {
+          let assetOrChunkFound: boolean = this.options.skipDetection || chunk.chunks.length > 0;
+          if (!assetOrChunkFound) {
+            chunk.forEachModule((innerModule: IV3Module) => {
+              if (innerModule.assets && Object.keys(innerModule.assets).length > 0) {
+                assetOrChunkFound = true;
+              }
+            });
+          }
+
+          if (assetOrChunkFound) {
+            return this._getStartupCode({
+              source,
+              chunk,
+              hash,
+              requireFn: compilation.mainTemplate.requireFn
+            });
+          } else {
+            return source;
+          }
+        });
+      });
+    }
+  }
+
+  private _getStartupCode(options: IStartupCodeOptions): string {
+    const moduleOptions: IInternalOptions = cloneDeep(this.options);
+
+    // If this module has ownership over any chunks or assets, inject the public path code
+    moduleOptions.webpackPublicPathVariable = `${options.requireFn}.p`;
+    moduleOptions.linePrefix = '  ';
+
+    if (this.options.scriptName) {
+      moduleOptions.regexName = this.options.scriptName.name;
+      if (this.options.scriptName.isTokenized) {
+        moduleOptions.regexName = moduleOptions.regexName
+          .replace(/\[name\]/g, escapeRegExp(options.chunk.name))
+          .replace(/\[hash\]/g, options.chunk.renderedHash);
+      }
+    }
+
+    return [
+      '// Set the webpack public path',
+      '(function () {',
+        getSetPublicPathCode(moduleOptions, console.error),
+      '})();',
+      '',
+      options.source
+    ].join(EOL);
   }
 }
