@@ -1,14 +1,12 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
- * See LICENSE in the project root for license information.
- */
+// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+// See LICENSE in the project root for license information.
 
 /**
  * This simple plugin sets the `__webpack_public_path__` variable to
  *  a value specified in the arguments, optionally appended to the SystemJs baseURL
  *  property.
+ * @packageDocumentation
  */
-declare const packageDescription: void; // tslint:disable-line:no-unused-variable
 
 export * from './SetPublicPathPlugin';
 export {
